0.4.3 Assets
Committing the Assets for 0.4.3

+++ Assets/Scripts/Pathfinder/AStarNode.cs
@@ -7,6 +7,6 @@ public class AStarNode
     public AStarNode parent;
     public Vector2 position;
     public int gCost;
-    public int hCost;
-    public int fCost { get { return gCost + hCost; } }
+    public float hCost;
+    public float fCost { get { return gCost + hCost; } }
 }
